backend/pyproject.toml
[tool.poetry]
name = "isochronic-backend"
version = "0.1.0"
description = "FastAPI backend for isochronic tone editor"
authors = ["Isochronic Team <team@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.111.0"
uvicorn = {version = "0.30.1", extras = ["standard"]}
sqlmodel = "0.0.22"
alembic = "1.13.1"
psycopg = {version = "3.1.19", extras = ["binary"]}
python-jose = {version = "3.3.0", extras = ["cryptography"]}
passlib = {version = "1.7.4", extras = ["bcrypt"]}
python-multipart = "0.0.9"
pydantic-settings = "2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.2"
httpx = "0.27.0"
ruff = "0.5.1"
mypy = "1.10.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "N", "ASYNC", "D"]
ignore = ["D100"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
